API

API

Types

SolidStateDetectors.ADLChargeDriftModel

SolidStateDetectors.CylindricalGrid

SolidStateDetectors.PointType

SolidStateDetectors.PointTypes

SolidStateDetectors.SolidStateDetector

SolidStateDetectors.VacuumChargeDriftModel

Functions

SolidStateDetectors.add_fano_noise

SolidStateDetectors.calculate_electric_potential

SolidStateDetectors.calculate_weighting_potential

SolidStateDetectors.extent_2D_grid_to_3D

SolidStateDetectors.extent_2D_grid_to_3D!

SolidStateDetectors.get_active_volume

SolidStateDetectors.get_path_to_example_config_files

SolidStateDetectors.read_from_hdf5

SolidStateDetectors.read_from_hdf5

SolidStateDetectors.write_to_hdf5

SolidStateDetectors.write_to_hdf5

Documentation

SolidStateDetectors.ADLChargeDriftModel — Type.

ADLChargeDriftModel{T <: AbstractFloat} <: AbstractChargeDriftModels

Fields

electrons::CarrierParameters{T}

holes::CarrierParameters{T}

`phi110::T

gammas::SVector{4, SMatrix{3,3,T}}

source

SolidStateDetectors.CylindricalGrid — Type.

CylindricalGrid{T<:AbstractFloat} <: Grid

Stores the electric potential on a three dimensional cylindrical grid and also the position of the grid points.

Fields

cyclic::T: Stores the periodicity of the system. E.g. 2π.

r::T: The r-axis of the grid in m.

φ::T: The φ-axis of the grid in rad.

z::T: The r-axis of the grid in m.

potential::Array{T, 3}: The potential values in V of each grid point.

source

SolidStateDetectors.PointTypes — Type.

PointTypes{T <: AbstractFloat} <: AbstractPointTypes

Stores the point type information in a three dimensional cylindrical grid and also the position of the grid points.

Fields

r::T: The r-axis of the grid in m.

φ::T: The φ-axis of the grid in rad.

z::T: The r-axis of the grid in m.

pointtypes::Array{PointType, 3}: The point type of each grid point.

source

SolidStateDetectors.SolidStateDetector — Method.

SolidStateDetector{T}(filename::AbstractString)::SolidStateDetector where T <: AbstractFloat

Reads in a config-JSON file and returns an Detector struct which holds all information specified in the config file.

source

SolidStateDetectors.VacuumChargeDriftModel — Type.

VacuumChargeDriftModel <: AbstractChargeDriftModels

source

SolidStateDetectors.add_fano_noise — Function.

add_fano_noise(E_dep::RealQuantity, E_ionisation::RealQuantity, f_fano::Real)::RealQuantity

Add Fano noise to an energy deposition E_dep, assuming a detector material ionisation energy E_ionisation and a Fano factor f_fano.

source

SolidStateDetectors.calculate_electric_potential — Method.

calculate_electric_potential(det::SolidStateDetector; <keyword arguments>)::Tuple{<:Grid, PointTypes}

Compute the electric potential for the given Detector det on an adaptive grid through successive over relaxation.

There are serveral <keyword arguments> which can be used to tune the computation:

Keywords

coordinates::Symbol: the kind of the coordinate system of the grid. Right now only :cylindrical is possible.

convergence_limit::Real: convergence_limit times the bias voltage sets the convergence limit of the relaxation. The convergence value is the absolute maximum difference of the potential between two iterations of all grid points. Default of convergence_limit is 5e-6 (times bias voltage).

max_refinements::Int: number of maximum refinements. Default is 2. Set it to 0 to switch off refinement.

refinement_limits::Vector{Real}: vector of refinement limits for each dimension (in case of cylindrical coordinates the order is r, φ, z). A refinement limit (e.g. refinement_limits[1]) times the bias voltage of the detector det is the maximum allowed voltage difference between two neighbouring grid points in the respective dimension. When the difference is larger, new points are created inbetween. Default is [1e-4, 1e-4, 1e-4].

min_grid_spacing::Vector{Real}: vector of the mimimum allowed distance between two grid points for each dimension. For normal coordinates the unit is meter. For angular coordinates, the unit is radiance. It prevents the refinement to make the grid to fine. Default is [1e-4, 1e-2, 1e-4].

depletion_handling::Bool: enables the handling of undepleted regions. Default is false.

nthreads::Int: Number of threads to use in the computation. Default is Base.Threads.nthreads(). The environment variable JULIA_NUM_THREADS must be set appropriately before the Julia session was started (e.g. export JULIA_NUM_THREADS=8 in case of bash).

sor_consts::Vector{<:Real}: Two element array. First element contains the SOR constant for r = 0. Second contains the constant at the outer most grid point in r. A linear scaling is applied in between. First element should be smaller than the second one and both should be ∈ [1.0, 2.0]. Default is [1.4, 1.85].

return_2π_grid::Bool: If set to true, the grid is extended to 2π in φ after it is computated. Default is true. This keyword is ignored if the simulation is in 2D. Use extent_2D_grid_to_3D() function to extend a 2D grid into 3D.

max_n_iterations::Int: Set the maximum number of iterations which are performed after each grid refinement. Default is 10000. If set to -1 there will be no limit.

verbose::Bool=true: Boolean whether info output is produced or not.

init_grid_spacing::Vector{<:Real}: Initial spacing of the grid. Default is [2e-3, 2π / 72, 2e-3] <=> [2mm, 5 degree, 2mm ]

Additional Information

The function always turns the detector into a p-type detector to compute the potential. At the end it turns the signs to make it n-type again if it was n-type.

source

SolidStateDetectors.calculate_weighting_potential — Method.

calculate_weighting_potential(det::SolidStateDetector, channels::Array{Int, 1}; <keyword arguments>)::Grid

Compute the weighting potential of the channels of the given Detector det on an adaptive grid through successive over relaxation. channels is a list of the channels which are fixed to 1. All other channels are fixed to 0.

There are serveral <keyword arguments> which can be used to tune the computation:

Keywords

coordinates::Symbol: the kind of the coordinate system of the grid. Right now only :cylindrical is possible.

convergence_limit::Real: convergence_limit times the bias voltage sets the convergence limit of the relaxation. The convergence value is the absolute maximum difference of the potential between two iterations of all grid points. Default of convergence_limit is 5e-6 (times bias voltage).

max_refinements::Int: number of maximum refinements. Default is 2. Set it to 0 to switch off refinement.

refinement_limits::Vector{Real}: vector of refinement limits for each dimension (in case of cylindrical coordinates the order is r, φ, z). A refinement limit (e.g. refinement_limits[1]) times the bias voltage of the detector det is the maximum allowed voltage difference between two neighbouring grid points in the respective dimension. When the difference is larger, new points are created inbetween. Default is [1e-4, 1e-4, 1e-4].

min_grid_spacing::Vector{Real}: vector of the mimimum allowed distance between two grid points for each dimension. For normal coordinates the unit is meter. For angular coordinates the unit is radiance. It prevents the refinement to make the grid to fine. Default is [1e-4, 1e-2, 1e-4].

depletion_handling::Bool: NOT IMPLEMENTET YET. Enables the handling of undepleted regions. Default is false.

nthreads::Int: Number of threads to use in the computation. Default is Base.Threads.nthreads(). The environment variable JULIA_NUM_THREADS must be set appropriately before the Julia session was started (e.g. export JULIA_NUM_THREADS=8 in case of bash).

sor_consts::Vector{<:Real}: Two element array. First element contains the SOR constant for r = 0. Second contains the constant at the outer most grid point in r. A linear scaling is applied in between. First element should be smaller than the second one and both should be ∈ [1.0, 2.0].

max_n_iterations::Int: Set the maximum number of iterations which are performed after each grid refinement. Default is 10000. If set to -1 there will be no limit.

verbose::Bool=true: Boolean whether info output is produced or not.

init_grid_spacing::Vector{<:Real}: Initial spacing of the grid. Default is [2e-3, 2π / 72, 2e-3] <=> [2mm, 5 degree, 2mm ]

Additional Information

source

SolidStateDetectors.get_active_volume — Method.

get_active_volume(grid::CylindricalGrid, pts::PointTypes{T}) where {T}

Returns an approximation of the active volume of the detector by summing up the cell volumes of all depleted cells.

source

SolidStateDetectors.PointType — Type.

const PointType = UInt8

Stores certain information about a grid point via bit-flags.

Right now there are:

`const update_bit      = 0x01`
`const undepleted_bit  = 0x02`
`const bubble_bit      = 0x04`
`const pn_junction_bit = 0x08`

How to get information out of a PointType variable pt:

pt & update_bit == 0 -> do not update this point (for fixed points)

pt & update_bit > 0 -> do update this point

pt & undepleted_bit > 0 -> this point is undepleted

pt & bubble_bit > 0 -> this point is in an "bubble"-area/volume. So an undepleted region which is not in contact with an electrode (fixed points).

pt & pn_junction_bit > 0 -> this point to the solid state detector. So it is in the volume of the n-type or p-type material.

source

SolidStateDetectors.extent_2D_grid_to_3D! — Method.

extent_2D_grid_to_3D!(grid::CylindricalGrid, n::Int)::Nothing

This function extend a 2-dimensional grid (only one tick in φ) to an 3-dimensional grid with n ticks in φ (symmetrically distributed up to 2π).

source

SolidStateDetectors.extent_2D_grid_to_3D — Method.

extent_2D_grid_to_3D(grid::CylindricalGrid, n::Int)::CylindricalGrid

This function returns a extended grid of a 2-dimensional grid grid. The extended grid has n ticks in φ (symmetrically distributed up to 2π).

source

SolidStateDetectors.get_path_to_example_config_files — Method.

get_path_to_example_config_files()::String

Returns the path to example config files provided by the package.

source

SolidStateDetectors.read_from_hdf5 — Method.

read_from_hdf5(input, ::Type{ElectricPotential})::ElectricPotential

input should be and HDF5.HDF5Group.

source

SolidStateDetectors.read_from_hdf5 — Method.

read_from_hdf5(input, ::Type{PointTypes})::PointTypes

input should be and HDF5.HDF5Group.

source

SolidStateDetectors.write_to_hdf5 — Method.

write_to_hdf5(output, ep::ElectricPotential)

output should be and HDF5.HDF5Group.

source

SolidStateDetectors.write_to_hdf5 — Method.

write_to_hdf5(output, ep::PointTypes)

output should be and HDF5.HDF5Group.

source