API

Dictionaries

SolidStateDetectors.SSD_examples

Types

SolidStateDetectors.ADLChargeDriftModel

SolidStateDetectors.AbstractChargeDensity

SolidStateDetectors.AbstractConfig

SolidStateDetectors.AbstractImpurityDensity

SolidStateDetectors.ConstantChargeDensity

SolidStateDetectors.ConstantImpurityDensity

SolidStateDetectors.ConstructiveSolidGeometry.Box

SolidStateDetectors.ConstructiveSolidGeometry.CSGDifference

SolidStateDetectors.ConstructiveSolidGeometry.CSGIntersection

SolidStateDetectors.ConstructiveSolidGeometry.CSGUnion

SolidStateDetectors.ConstructiveSolidGeometry.CartesianPoint

SolidStateDetectors.ConstructiveSolidGeometry.CartesianVector

SolidStateDetectors.ConstructiveSolidGeometry.Cone

SolidStateDetectors.ConstructiveSolidGeometry.ConeMantle

SolidStateDetectors.ConstructiveSolidGeometry.CylindricalPoint

SolidStateDetectors.ConstructiveSolidGeometry.CylindricalVector

SolidStateDetectors.ConstructiveSolidGeometry.Ellipse

SolidStateDetectors.ConstructiveSolidGeometry.EllipticalSurface

SolidStateDetectors.Contact

SolidStateDetectors.CylindricalChargeDensity

SolidStateDetectors.CylindricalImpurityDensity

SolidStateDetectors.DielectricDistribution

SolidStateDetectors.DiscreteAxis

SolidStateDetectors.DiscreteAxis

SolidStateDetectors.EffectiveChargeDensity

SolidStateDetectors.ElectricFieldChargeDriftModel

SolidStateDetectors.ElectricPotential

SolidStateDetectors.Event

SolidStateDetectors.Grid

SolidStateDetectors.Grid

SolidStateDetectors.Grid

SolidStateDetectors.LinearChargeDensity

SolidStateDetectors.LinearImpurityDensity

SolidStateDetectors.PointType

SolidStateDetectors.PointTypes

SolidStateDetectors.Simulation

SolidStateDetectors.SolidStateDetector

SolidStateDetectors.WeightingPotential

Functions

SolidStateDetectors.ConstructiveSolidGeometry.extreme_points

SolidStateDetectors.ConstructiveSolidGeometry.intersection

SolidStateDetectors.ConstructiveSolidGeometry.intersection

SolidStateDetectors.ConstructiveSolidGeometry.intersection

SolidStateDetectors.ConstructiveSolidGeometry.intersection

SolidStateDetectors._convert_to_original_potential

SolidStateDetectors._drift_charge!

SolidStateDetectors._get_closed_potential

SolidStateDetectors.add_fano_noise

SolidStateDetectors.apply_initial_state!

SolidStateDetectors.apply_initial_state!

SolidStateDetectors.calculate_capacitance

SolidStateDetectors.calculate_drift_fields!

SolidStateDetectors.calculate_electric_field!

SolidStateDetectors.calculate_electric_potential!

SolidStateDetectors.calculate_weighting_potential!

SolidStateDetectors.drift_charges!

SolidStateDetectors.get_active_volume

SolidStateDetectors.get_electron_drift_field

SolidStateDetectors.get_hole_drift_field

SolidStateDetectors.get_path_to_example_config_files

SolidStateDetectors.parse_config_file

SolidStateDetectors.readsiggen

SolidStateDetectors.refine!

SolidStateDetectors.refine!

SolidStateDetectors.refine_scalar_potential

SolidStateDetectors.siggentodict

SolidStateDetectors.simulate!

SolidStateDetectors.simulate!

SolidStateDetectors.simulate_waveforms

SolidStateDetectors.update_till_convergence!

SolidStateDetectors.update_till_convergence!

Documentation

SolidStateDetectors.SSD_examples — Constant

SSD_examples::Dict{Symbol,String}

Dictionary with the paths to the example detector configuration files provided by the package.

Find the possible keys of the dictionary with keys(SSD_examples).

The example detector configuration files can be loaded via

path_to_config_file = SSD_examples[:InvertedCoax]
sim = Simulation(path_to_config_file)

SolidStateDetectors.ADLChargeDriftModel — Type

ADLChargeDriftModel{T <: SSDFloat, M <: AbstractDriftMaterial, N, TM <: AbstractTemperatureModel{T}} <: AbstractChargeDriftModel{T}

Fields

electrons::CarrierParameters{T}

holes::CarrierParameters{T}

crystal_orientation::SMatrix{3,3,T,9}

γ::SVector{N,SMatrix{3,3,T,9}}

parameters::ADLParameters{T}

temperaturemodel::TM

SolidStateDetectors.AbstractChargeDensity — Type

# abstract type AbstractChargeDensity{T <: SSDFloat} end

Charge densities <: AbstractChargeDensity should be defined to return (via get_charge_density-method) a charge density in SI units, thus, in Q/m^3.

SolidStateDetectors.AbstractConfig — Type

abstract type AbstractConfig{T <: SSDFloat} end

Supertype of all detector/world/object configs.

User defined geometries must be subtype of AbstractConfig{T}.

There are a few functions which must be defined for a user config, e.g. struct UserConfig{T} <: AbstractConfig{T}:

For cylindrical grids:

in(pt::CylindricalPoint{T}, config::UserConfig{T})::Bool where {T <: SSDFloat}

Grid(config::UserConfig{T})::Grid{T, 3, Cylindrical} where {T <: SSDFloat}

get_ρ_and_ϵ(pt::CylindricalPoint{T}, config::UserConfig{T})::Tuple{T, T} where {T <: SSDFloat}

set_pointtypes_and_fixed_potentials!(pointtypes::Array{PointType, 3}, potential::Array{T, 3}, grid::Grid{T, 3, Cylindrical}, config::UserConfig{T}; weighting_potential_contact_id::Union{Missing, Int} = missing)::Nothing where {T <: SSDFloat}

For cartesian grids:

in(pt::CartesianPoint{3, T}, config::UserConfig{T})::Bool

Grid(config::UserConfig{T})::Grid{T, 3, Cartesian} where {T <: SSDFloat}

get_ρ_and_ϵ(pt::CartesianPoint{3, T}, config::UserConfig{T})::Tuple{T, T} where {T <: SSDFloat}

set_pointtypes_and_fixed_potentials!(pointtypes::Array{PointType, 3}, potential::Array{T, 3}, grid::Grid{T, 3, Cartesian}, config::UserConfig{T}; weighting_potential_contact_id::Union{Missing, Int} = missing)::Nothing where {T <: SSDFloat}

SolidStateDetectors.AbstractImpurityDensity — Type

# abstract type AbstractImpurityDensity{T <: SSDFloat} end

Impurity densities <: AbstractImpurityDensity should be defined to return (via get_impurity_density-method) a density of impurities in SI units, thus, in 1/m^3. For semiconductors, this will be converted internally into a charge distribution by multiplication with the elementary charge. The sign of the impurity density is important. It is taken into account in the conversion to a charge density and, thus, defines where the semiconductor is n-type or p-type.

SolidStateDetectors.ConstantChargeDensity — Type

struct ConstantChargeDensity{T <: SSDFloat} <: AbstractChargeDensity{T}

Returns always a fixed charge density.

SolidStateDetectors.ConstantImpurityDensity — Type

struct ConstantImpurityDensity{T <: SSDFloat} <: AbstractImpurityDensity{T}

Returns always a fixed impurity density.

SolidStateDetectors.Contact — Type

mutable struct Contact{T} <: AbstractContact{T}

T: Type of precision.

SolidStateDetectors.CylindricalChargeDensity — Type

struct CylindricalChargeDensity{T <: SSDFloat} <: AbstractChargeDensity{T}

Simple charge density model which assumes a linear gradient in charge density in each spatial dimension of a cylindrical coordinate system. offsets::NTuple{3, T} are the charge densities at 0 and gradients::NTuple{3, T} are the linear slopes in r and z direction.

SolidStateDetectors.CylindricalImpurityDensity — Type

struct CylindricalImpurityDensity{T <: SSDFloat} <: AbstractImpurityDensity{T}

Simple Impurity density model which assumes a linear gradient in Impurity density in each spatial dimension of a cylindrical coordinate system. offsets::NTuple{3, T} are the Impurity densities at 0 and gradients::NTuple{3, T} are the linear slopes in r and z direction.

SolidStateDetectors.DielectricDistribution — Type

struct DielectricDistribution{T, N, S, AT} <: AbstractArray{T, N}

T: Element type of data.

N: Dimension of the grid and data array.

S: Coordinate system (Cartesian or Cylindrical).

AT: Axes type.

Fields

data::Array{T, N}

grid::Grid{T, N, S, AT}

The data array contains the values of the dielectric distribution at the discrete points between the points defined by the axes ticks of the extended grid of grid.

Thus, size(data) == size(grid) .+ 1 !

The unit of the values in data is the unit one ([1]).

SolidStateDetectors.DiscreteAxis — Type

DiscreteAxis{T, BL, BR, I} <: AbstractAxis{T, BL, BR, I}

T: Type of ticks

BL: left boundary condition

BR: right boundary condition

BL, BR ∈ {:periodic, :reflecting, :infinite, :r0, :fixed}

I: IntervalSets.Interval (closed or open boundaries)

SolidStateDetectors.DiscreteAxis — Method

DiscreteAxis(left_endpoint::T, right_endpoint::T, BL::Symbol, BR::Symbol, L::Symbol, R::Symbol, ticks::AbstractVector{T}) where {T}

T: Type of ticks

BL, BR ∈ {:periodic, :reflecting, :infinite, :r0, :fixed}

L, R ∈ {:closed, :open}

ticks: Ticks of the axis

SolidStateDetectors.EffectiveChargeDensity — Type

struct EffectiveChargeDensity{T, N, S, AT} <: AbstractArray{T, N}

T: Element type of data.

N: Dimension of the grid and data array.

S: Coordinate system (Cartesian or Cylindrical).

AT: Axes type.

Fields

data::Array{T, N}

grid::Grid{T, N, S, AT}

The data array contains the values of the effective charge density at the discrete points defined by the axes ticks of the grid.

The effective charge density is the charge density ([Q/m^3]) times the volume of the voxel of the respective grid point ([m^3]). Thus, the units of the effective charge density in data is coulombs ([C]).

SolidStateDetectors.ElectricFieldChargeDriftModel — Type

ElectricFieldChargeDriftModel <: AbstractChargeDriftModel

SolidStateDetectors.ElectricPotential — Type

struct ElectricPotential{T, N, S, AT} <: AbstractArray{T, N}

T: Element type of data.

N: Dimension of the grid and data array.

S: Coordinate system (Cartesian or Cylindrical).

AT: Axes type.

Fields

data::Array{T, N}

grid::Grid{T, N, S, AT}

The data array contains the values of the electric potential at the discrete points defined by the axes ticks of the grid.

The unit of the values in data is volts ([V]).

SolidStateDetectors.Event — Type

mutable struct Event{T <: SSDFloat}

Collection struct for individual events. This (mutable) struct is mend to be used to look at individual events, not to process a huge amount of events.

Fields:

locations::Union{Vector{<:AbstractCoordinatePoint{T}}, Missing}: Vector of the positions of all hits of the event.

energies::Union{Vector{T}, Missing}: Vector of energies corresponding to the hits of the event.

drift_paths::Union{Vector{EHDriftPath{T}}, Missing}: Calculated drift paths of each hit position.

waveforms::Union{Vector{<:Any}, Missing}: Generated signals (waveforms) of the event.

SolidStateDetectors.Grid — Type

struct Grid{T, N, S <: AbstractCoordinateSystem, AT} <: AbstractGrid{T, N}

T: Tick type (element type) of the axes.

N: Dimension of the grid.

S: Coordinate system (Cartesian or Cylindrical).

AT: Axes type.

Fields

axes::AT: Tuple of length N containing DiscreteAxis.

SolidStateDetectors.Grid — Method

Grid(sim::Simulation{T, Cartesian})

Initialize a grid based on the sim::Simulation{T, S}.

The grid can be tuned via the two keywords max_tick_distance and max_distance_ratio::Real = 5.

max_tick_distance can either be a Quantity, e.g. 1u"mm", or a Tuple of Quantities, e.g. (1u"mm", 0.2u"cm", 3u"mm"), to set it for each axis of the Grid separately. If max_tick_distance is missing, one fourth of the axis length is used.

Keywords:

for_weighting_potential::Bool = false

max_tick_distance::Union{Missing, LengthQuantity, Tuple{LengthQuantity, LengthQuantity, LengthQuantity}}} = missing

max_distance_ratio::Real = 5

full_2π::Bool = false

SolidStateDetectors.Grid — Method

Grid(sim::Simulation{T, Cylindrical})

Initialize a grid based on the sim::Simulation{T, S}.

The grid can be tuned via the two keywords max_tick_distance and max_distance_ratio::Real = 5.

max_tick_distance can either be a Quantity, e.g. 1u"mm", or a Tuple of Quantities, e.g. (1u"mm", 15u"°", 3u"mm"), to set it for each axis of the Grid separately. If max_tick_distance is missing, one fourth of the axis length is used.

Keywords:

for_weighting_potential::Bool = false

max_tick_distance::Union{Missing, LengthQuantity, Tuple{LengthQuantity, AngleQuantity, LengthQuantity}}} = missing

max_distance_ratio::Real = 5

full_2π::Bool = false

SolidStateDetectors.LinearChargeDensity — Type

struct LinearChargeDensity{T <: SSDFloat} <: AbstractChargeDensity{T}

Simple charge density model which assumes a linear gradient in charge density in each dimension of a Cartesian coordinate system. offsets::NTuple{3, T} are the charge densities at 0 and gradients::NTuple{3, T} are the linear slopes in each dimension, x, y and z.

SolidStateDetectors.LinearImpurityDensity — Type

struct LinearImpurityDensity{T <: SSDFloat} <: AbstractImpurityDensity{T}

Simple Impurity density model which assumes a linear gradient in Impurity density in each dimension of a Cartesian coordinate system. offsets::NTuple{3, T} are the Impurity densities at 0 and gradients::NTuple{3, T} are the linear slopes in each dimension, x, y and z.

SolidStateDetectors.PointType — Type

const PointType = UInt8

Stores certain information about a grid point via bit-flags.

Right now there are:

`const update_bit      = 0x01`
`const undepleted_bit  = 0x02`
`const pn_junction_bit = 0x04`

How to get information out of a PointType variable pt:

pt & update_bit == 0 -> do not update this point (for fixed points)

pt & update_bit > 0 -> do update this point

pt & undepleted_bit > 0 -> this point is undepleted

pt & pn_junction_bit > 0 -> this point belongs to the solid state detector. So it is in the volume of the n-type or p-type material.

SolidStateDetectors.PointTypes — Type

struct PointTypes{T, N, S, AT} <: AbstractArray{T, N}

T: Element type of grid.axes.

N: Dimension of the grid and data array.

S: Coordinate system (Cartesian or Cylindrical).

AT: Axes type.

Fields

data::Array{PointType, N}

grid::Grid{T, N, S, AT}

The data array contains the point types of the discrete points defined by the axes ticks of the grid.

The element type data is PointType which is an UInt8 and stores is information bitwise, see PointType.

SolidStateDetectors.Simulation — Type

mutable struct Simulation{T <: SSDFloat, CS <: AbstractCoordinateSystem} <: AbstractSimulation{T}

Collection of all parts of a Simulation of a Solid State Detector.

SolidStateDetectors.SolidStateDetector — Type

struct SolidStateDetector{T,SC,CT,PT,VDM} <: AbstractConfig{T}

Fields

name::String: Name of the detector

semiconductor::SC: Semiconductor of the detector.

contacts::CT: Vector of Contacts of the detector.

passives::PT: Vector of Passive objects, e.g. holding structures around the detector.

virtual_drift_volumes::VDM: Vector of virtual drift volumes in which the drift can be modulated by an user-defined method for modulate_driftvector.

SolidStateDetectors.WeightingPotential — Type

struct WeightingPotential{T, N, S, AT} <: AbstractArray{T, N}

T: Element type of data.

N: Dimension of the grid and data array.

S: Coordinate system (Cartesian or Cylindrical).

AT: Axes type.

Fields

data::Array{T, N}

grid::Grid{T, N, S, AT}

The data array contains the values of the weighting potential at the discrete points defined by the axes ticks of the grid.

The unit of the values in data is the unit one ([1]).

SolidStateDetectors._convert_to_original_potential — Method

_convert_to_original_potential(::Type{P}, data, grid) where {P, T, CS}

Basically the counterpart to _get_closed_potential.

SolidStateDetectors._drift_charge! — Method

_drift_charge!(...)

Before calling this function one should check that startpos is inside det: in(startpos, det)

SolidStateDetectors._get_closed_potential — Method

_get_closed_potential(p::ScalarPotential{T,3,CS}) where {T, CS}

Returns an closed Grid & Potential: E.g. if one of the axis is {:closed,:open} it will turn this into {:closed,:closed} and also extend the data field of the potential in the respective dimension and fill it with the respective values.

SolidStateDetectors.add_fano_noise — Function

add_fano_noise(E_dep::RealQuantity, E_ionisation::RealQuantity, f_fano::Real)::RealQuantity

Add Fano noise to an energy deposition E_dep, assuming a detector material ionisation energy E_ionisation and a Fano factor f_fano.

SolidStateDetectors.apply_initial_state! — Method

apply_initial_state!(sim::Simulation{T}, ::Type{ElectricPotential}, grid::Grid{T} = Grid(sim))::Nothing

Applies the initial state of the electric potential calculation. It overwrites sim.electric_potential, sim.q_eff_imp, sim.q_eff_fix, sim.ϵ and sim.point_types.

SolidStateDetectors.apply_initial_state! — Method

apply_initial_state!(sim::Simulation{T}, ::Type{WeightingPotential}, contact_id::Int, grid::Grid{T} = Grid(sim))::Nothing

Applies the initial state of the weighting potential calculation for the contact with the id contact_id. It overwrites sim.weighting_potentials[contact_id].

SolidStateDetectors.calculate_capacitance — Method

calculate_capacitance(sim::Simulation{T})::T where {T <: SSDFloat}

Calculates the capacitance of an detector in Farad.

SolidStateDetectors.calculate_drift_fields! — Method

calculate_drift_fields!(sim::Simulation{T}; use_nthreads::Int = Base.Threads.nthreads())

Calculates the drift fields for electrons and holes and stores them in sim.sim.electron_drift_field and sim.hole_drift_field. The fields are calculated from the electric field, sim.electric_field, and the drift model sim.detector.semiconductor.charge_drift_model.

SolidStateDetectors.calculate_electric_field! — Method

calculate_electric_field!(sim::Simulation{T}, args...; n_points_in_φ::Union{Missing, Int} = missing, kwargs...)::Nothing

Calculates the electric field from the electric potential stored in sim.electric_potential and stores is under sim.electric_field.

Keywords

n_points_in_φ: If the electric potential is 2D (cylindrical coordinates and symmetric in φ), the electric potential

is extended to n_points_in_φ "layers" in φ in order to calculate a 3D electric field.

SolidStateDetectors.calculate_electric_potential! — Method

calculate_electric_potential!(sim::Simulation{T}; kwargs...)::Nothing

Compute the electric potential for the given Simulation sim on an adaptive grid through successive over relaxation.

There are several <keyword arguments> which can be used to tune the computation:

Keywords

convergence_limit::Real: convergence_limit times the bias voltage sets the convergence limit of the relaxation. The convergence value is the absolute maximum difference of the potential between two iterations of all grid points. Default of convergence_limit is 2e-6 (times bias voltage).

refinement_limits: Defines the maximum relative (to applied bias voltage) allowed differences of the potential value of neighbored grid points in each dimension for each refinement.

rl::Real -> One refinement with rl equal in all 3 dimensions

rl::Tuple{<:Real,<:Real,<:Real} -> One refinement with rl set individual for each dimension

rl::Vector{<:Real} -> length(l) refinements with rl[i] being the limit for the i-th refinement.

rl::Vector{<:Real,<:Real,<:Real}} -> length(rl) refinements with rl[i] being the limits for the i-th refinement.

min_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the minimum allowed distance between two grid ticks for each dimension. It prevents the refinement to make the grid to fine. Default is 1e-5 for linear axes and 1e-5 / (0.25 * r_max) for the polar axis in case of a cylindrical grid.

max_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the maximum allowed distance between two grid ticks for each dimension used in the initialization of the grid. Default is 1/4 of size of the world of the respective dimension.

grid::Grid: Initial grid used to start the simulation. Default is Grid(sim).

max_distance_ratio::Real: Maximum allowed ratio between the two distances in any dimension to the two neighbouring grid points. If the ratio is too large, additional ticks are generated such that the new ratios are smaller than max_distance_ratio. Default is 5.

depletion_handling::Bool: Enables the handling of undepleted regions. Default is false.

use_nthreads::Int: Number of threads to use in the computation. Default is Base.Threads.nthreads(). The environment variable JULIA_NUM_THREADS must be set appropriately before the Julia session was started (e.g. export JULIA_NUM_THREADS=8 in case of bash).

sor_consts::Union{<:Real, NTuple{2, <:Real}}: Two element tuple in case of cylindrical coordinates. First element contains the SOR constant for r = 0. Second contains the constant at the outer most grid point in r. A linear scaling is applied in between. First element should be smaller than the second one and both should be ∈ [1.0, 2.0]. Default is [1.4, 1.85]. In case of cartesian coordinates only one value is taken.

max_n_iterations::Int: Set the maximum number of iterations which are performed after each grid refinement. Default is 10000. If set to -1 there will be no limit.

not_only_paint_contacts::Bool = true: Whether to only use the painting algorithm of the contacts. Setting it to false should improve the performance but the inside of contacts are not fixed points anymore.

paint_contacts::Bool = true: Enable or disable the paining of the surfaces of the contacts onto the grid.

verbose::Bool=true: Boolean whether info output is produced or not.

SolidStateDetectors.calculate_weighting_potential! — Method

calculate_weighting_potential!(sim::Simulation{T}, contact_id::Int; kwargs...)::Nothing

Compute the weighting potential for the contact with id contact_id for the given Simulation sim on an adaptive grid through successive over relaxation.

There are several <keyword arguments> which can be used to tune the computation:

Keywords

convergence_limit::Real: convergence_limit times the bias voltage sets the convergence limit of the relaxation. The convergence value is the absolute maximum difference of the potential between two iterations of all grid points. Default of convergence_limit is 2e-6 (times bias voltage).

refinement_limits: Defines the maximum relative (to applied bias voltage) allowed differences of the potential value of neighbored grid points in each dimension for each refinement.

rl::Real -> One refinement with rl equal in all 3 dimensions

rl::Tuple{<:Real,<:Real,<:Real} -> One refinement with rl set individual for each dimension

rl::Vector{<:Real} -> length(l) refinements with rl[i] being the limit for the i-th refinement.

rl::Vector{<:Real,<:Real,<:Real}} -> length(rl) refinements with rl[i] being the limits for the i-th refinement.

min_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the minimum allowed distance between two grid ticks for each dimension. It prevents the refinement to make the grid to fine. Default is 1e-5 for linear axes and 1e-5 / (0.25 * r_max) for the polar axis in case of a cylindrical grid.

max_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the maximum allowed distance between two grid ticks for each dimension used in the initialization of the grid. Default is 1/4 of size of the world of the respective dimension.

grid::Grid: Initial grid used to start the simulation. Default is Grid(sim).

max_distance_ratio::Real: Maximum allowed ratio between the two distances in any dimension to the two neighbouring grid points. If the ratio is too large, additional ticks are generated such that the new ratios are smaller than max_distance_ratio. Default is 5.

depletion_handling::Bool: Enables the handling of undepleted regions. Default is false.

use_nthreads::Int: Number of threads to use in the computation. Default is Base.Threads.nthreads(). The environment variable JULIA_NUM_THREADS must be set appropriately before the Julia session was started (e.g. export JULIA_NUM_THREADS=8 in case of bash).

sor_consts::Union{<:Real, NTuple{2, <:Real}}: Two element tuple in case of cylindrical coordinates. First element contains the SOR constant for r = 0. Second contains the constant at the outer most grid point in r. A linear scaling is applied in between. First element should be smaller than the second one and both should be ∈ [1.0, 2.0]. Default is [1.4, 1.85]. In case of cartesian coordinates only one value is taken.

max_n_iterations::Int: Set the maximum number of iterations which are performed after each grid refinement. Default is 10000. If set to -1 there will be no limit.

not_only_paint_contacts::Bool = true: Whether to only use the painting algorithm of the contacts. Setting it to false should improve the performance but the inside of contacts are not fixed points anymore.

paint_contacts::Bool = true: Enable or disable the paining of the surfaces of the contacts onto the grid.

verbose::Bool=true: Boolean whether info output is produced or not.

SolidStateDetectors.drift_charges! — Method

drift_charges!(event::Event{T}, sim::Simulation{T}; max_nsteps::Int = 1000, Δt::RealQuantity = 5u"ns", verbose::Bool = true)::Nothing where {T <: SSDFloat}

Calculate the drift paths for the given event and Simulation sim and stores them under event.drift_paths.

Keywords

Δt::RealQuantity = 5u"ns": Time difference between two time stamps of the drift.

max_nsteps::Int = 1000: Maximum number of steps in the drift of each hit.

verbose = false: Activate or deactivate additional info output.

SolidStateDetectors.get_active_volume — Method

get_active_volume(pts::PointTypes{T}) where {T}

Returns an approximation of the active volume of the detector by summing up the cell volumes of all depleted cells.

SolidStateDetectors.get_electron_drift_field — Method

get_electron_drift_field(ef::Array{SVector{3, T},3}, chargedriftmodel::AbstractChargeDriftModel)::Array{SVector{3,T},3} where {T <: SSDFloat}

Applies the charge drift model onto the electric field vectors. The field vectors have to be in cartesian coordinates.

SolidStateDetectors.get_hole_drift_field — Method

get_hole_drift_field(ef::Array{SVector{3, T},3}, chargedriftmodel::AbstractChargeDriftModel)::Array{SVector{3,T},3} where {T <: SSDFloat}

Applies the charge drift model onto the hole field vectors. The field vectors have to be in cartesian coordinates.

SolidStateDetectors.get_path_to_example_config_files — Method

get_path_to_example_config_files()::String

Returns the path to the example detector configuration files provided by the package.

SolidStateDetectors.parse_config_file — Method

parse_config_file(filename::AbstractString)::SolidStateDetector{T} where {T <: SSDFloat}

Reads in a config file and returns an Detector struct which holds all information specified in the config file. Currently supported formats for the config file: .json, .yaml

SolidStateDetectors.readsiggen — Method

readsiggen(file_path::String[, T::Type=Float64])

Read the '*.config' file in 'file_path' for SigGen and returns a dictionary of all parameters. Non-existing parameteres are set to 0. ...

Arguments

file_path::String: file path for the SigGen config file.

T::Type=Float64: type of the parameters in the output dictionary.

...

SolidStateDetectors.refine! — Method

refine!(sim::Simulation{T}, ::Type{ElectricPotential}, max_diffs::Tuple{<:Real,<:Real,<:Real}, minimum_distances::Tuple{<:Real,<:Real,<:Real})

Takes the current state of sim.electric_potential and refines it with respect to the input arguments max_diffs and minimum_distances.

SolidStateDetectors.refine! — Method

refine!(sim::Simulation{T}, ::Type{WeightingPotential}, max_diffs::Tuple{<:Real,<:Real,<:Real}, minimum_distances::Tuple{<:Real,<:Real,<:Real})

Takes the current state of sim.weighting_potentials[contact_id] and refines it with respect to the input arguments max_diffs and minimum_distances.

SolidStateDetectors.refine_scalar_potential — Method

refine_scalar_potential(p::ScalarPotential{T}, max_diffs::NTuple{3, T}, minimum_distances::NTuple{3, T}; 
    only2d::Val{only_2d} = Val(size(p.data, 2)==1)) where {T, only_2d}

Refine any scalar potential p.

Extent the grid to be a closed grid in all dimensions.

Refine the axis of the grid based on max_diffs and minimum_applied_potential: Insert N new ticks between to existing ticks such that the potential difference between each tick becomes smaller than max_diff[i] (i -> dimension) but that the distances between the ticks stays larger than minimum_distances[i].

Create the new data array for the refined grid and fill it by interpolation of the the initial (coarse) grid.

SolidStateDetectors.siggentodict — Method

siggentodict(config::Dict[, units::Dict, detector_name::String])

Converts the dictionary containing the parameters from a SigGen config file to a SSD config dictionary. This dictionary can be saved as a JSON file using the JSON package and 'JSON.print(file, config, 4)'. The 'detector_name' is set to "Public Inverted Coax" by default to inherit the colour scheme. ...

Arguments

config::Dict: dictionary containing SigGen parameters (output of readsiggen()).

units::Dict: units used in SigGen file (set to 'mm', 'deg', 'V' and 'K').

detector_name::String: name of the detector.

...

SolidStateDetectors.simulate! — Method

simulate!( sim::Simulation{T};  
                refinement_limits = [0.2, 0.1, 0.05],
                min_tick_distance::Union{Missing, LengthQuantity, Tuple{LengthQuantity, AngleQuantity, LengthQuantity}} = missing,
                max_tick_distance::Union{Missing, LengthQuantity, Tuple{LengthQuantity, AngleQuantity, LengthQuantity}} = missing,
                max_distance_ratio::Real = 5,
                verbose::Bool = false,
                use_nthreads::Int = Base.Threads.nthreads(),
                sor_consts::Union{Missing, <:Real, Tuple{<:Real,<:Real}} = missing,
                max_n_iterations::Int = -1,
                depletion_handling::Bool = false, 
                convergence_limit::Real = 1e-7 ) where {T <: SSDFloat}

Keywords

convergence_limit::Real: convergence_limit times the bias voltage sets the convergence limit of the relaxation. The convergence value is the absolute maximum difference of the potential between two iterations of all grid points. Default of convergence_limit is 2e-6 (times bias voltage).

refinement_limits: Defines the maximum relative (to applied bias voltage) allowed differences of the potential value of neighbored grid points in each dimension for each refinement.

rl::Real -> One refinement with rl equal in all 3 dimensions

rl::Tuple{<:Real,<:Real,<:Real} -> One refinement with rl set individual for each dimension

rl::Vector{<:Real} -> length(l) refinements with rl[i] being the limit for the i-th refinement.

rl::Vector{<:Real,<:Real,<:Real}} -> length(rl) refinements with rl[i] being the limits for the i-th refinement.

min_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the minimum allowed distance between two grid ticks for each dimension. It prevents the refinement to make the grid to fine. Default is 1e-5 for linear axes and 1e-5 / (0.25 * r_max) for the polar axis in case of a cylindrical grid.

max_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the maximum allowed distance between two grid ticks for each dimension used in the initialization of the grid. Default is 1/4 of size of the world of the respective dimension.

max_distance_ratio::Real: Maximum allowed ratio between the two distances in any dimension to the two neighbouring grid points. If the ratio is too large, additional ticks are generated such that the new ratios are smaller than max_distance_ratio. Default is 5.

depletion_handling::Bool: Enables the handling of undepleted regions. Default is false.

use_nthreads::Int: Number of threads to use in the computation. Default is Base.Threads.nthreads(). The environment variable JULIA_NUM_THREADS must be set appropriately before the Julia session was started (e.g. export JULIA_NUM_THREADS=8 in case of bash).

sor_consts::Union{<:Real, NTuple{2, <:Real}}: Two element tuple in case of cylindrical coordinates. First element contains the SOR constant for r = 0. Second contains the constant at the outer most grid point in r. A linear scaling is applied in between. First element should be smaller than the second one and both should be ∈ [1.0, 2.0]. Default is [1.4, 1.85]. In case of cartesian coordinates only one value is taken.

max_n_iterations::Int: Set the maximum number of iterations which are performed after each grid refinement. Default is -1. If set to -1 there will be no limit.

not_only_paint_contacts::Bool = true: Whether to only use the painting algorithm of the contacts. Setting it to false should improve the performance but the inside of contacts are not fixed points anymore.

paint_contacts::Bool = true: Enable or disable the paining of the surfaces of the contacts onto the grid.

verbose::Bool=true: Boolean whether info output is produced or not.

SolidStateDetectors.simulate! — Method

simulate!(event::Event{T}, sim::Simulation{T}; max_nsteps::Int = 1000, Δt::RealQuantity = 5u"ns", verbose::Bool = true)::Nothing where {T <: SSDFloat}

Simulates the event for a given Simulation sim: 1. Simulate the drift paths of all hits, at event.locations, and stores them under event.drift_paths. 2. Generate the signal (waveform) for all channels for which a weighting potential is specified in sim.weighting_potentials and stores them under event.waveforms.

Keywords

Δt::RealQuantity = 5u"ns": Time difference between two time stamps of the drift and the signals.

max_nsteps::Int = 1000: Maximum number of steps in the drift of each hit.

verbose = false: Activate or deactivate additional info output.

SolidStateDetectors.simulate_waveforms — Method

simulate_waveforms( mcevents::TypedTables.Table, s::Simulation{T})

Calculates the drift paths of all energy hits defined in mcevents based on the drift fields for electrons and holes stored in s.electron_drift_field and s.hole_drift_field.

Determines the signal (waveforms) of all channels (for which a weighting potential is given in the simulation object s.) for the generated drift paths of the hits.

Returns a the input table mcevents with an additional column waveform in which the generated waveforms are stored.

Note: The drift paths are just calculated temporarily and not returned.

Keywords

Δt::RealQuantity = 4u"ns": Time difference between two time stamps of the drift and the signals.

max_nsteps::Int = 1000: Maximum number of steps in the drift of each hit.

verbose = false: Activate or deactivate additional info output.

SolidStateDetectors.update_till_convergence! — Method

update_till_convergence!( sim::Simulation{T} ::Type{ElectricPotential}, convergence_limit::Real; kwargs...)::T

Takes the current state of sim.electric_potential and updates it until it has converged.

SolidStateDetectors.update_till_convergence! — Method

update_till_convergence!( sim::Simulation{T} ::Type{WeightingPotential}, contact_id::Int, convergence_limit::Real; kwargs...)::T

Takes the current state of sim.weighting_potentials[contact_id] and updates it until it has converged.

SolidStateDetectors.ConstructiveSolidGeometry.Box — Type

Box{T, CO} <: AbstractVolumePrimitive{T}

T: Type of values, e.g. Float64 CO: ClosedPrimitive or OpenPrimitive <-> whether surface belongs to it or not

SolidStateDetectors.ConstructiveSolidGeometry.CSGDifference — Type

struct CSGDifference{T, A <: AbstractGeometry{T}, B <: AbstractGeometry{T}} <: AbstractConstructiveGeometry{T}

a && !b

SolidStateDetectors.ConstructiveSolidGeometry.CSGIntersection — Type

struct CSGIntersection{T, A <: AbstractGeometry{T}, B <: AbstractGeometry{T}} <: AbstractConstructiveGeometry{T}

a && b

SolidStateDetectors.ConstructiveSolidGeometry.CSGUnion — Type

struct CSGUnion{T, A <: AbstractGeometry{T}, B <: AbstractGeometry{T}} <: AbstractConstructiveGeometry{T}

a || b

SolidStateDetectors.ConstructiveSolidGeometry.CartesianPoint — Type

struct CartesianPoint{T} <: AbstractCoordinatePoint{T, Cartesian}

x: x-coordinate in meter

y: y-coordinate in meter

z: z-coordinate in meter

SolidStateDetectors.ConstructiveSolidGeometry.CartesianVector — Type

struct CartesianVector{T} <: AbstractCoordinateVector{T, Cartesian}

x: x-component in meter

y: y-component in meter

z: z-component in meter

SolidStateDetectors.ConstructiveSolidGeometry.Cone — Type

struct Cone{T,CO,TR,TP} <: AbstractVolumePrimitive{T, CO}

T: Type of values, e.g. Float64 CO: ClosedPrimitive or OpenPrimitive <-> whether surface belongs to it or not

r::TR:

TR = Real -> Cylinder

TR = (Real, Real) -> Tube (rin = r[1], rout = r[2])

TR = ((Real,), (Real,)) Solid widening Cylinder -> (rbot = r[1][1], rtop = r[1][2])

TR = ((Real,Real), (Real,Real)) Solid widening Tube ->

(rbotin = r[1][1], rbotout = r[1][2], rtopin = r[2][1], rtopout = r[2][2]) * TR = (Nothing, (Real,Real)) Cone -> (rbotin = rbotout = 0, rtopin = r[2][1], rtopout = r[2][2]) * TR = ((Real,Real), Nothing) Cone -> (rbotin = r[1][1], rbotout = r[1][2], rtopin = rtopout = 0) * ... (+ elliptical cases -> (a, b)) * Not all are implemented yet

φ::TP:

TP = Nothing <-> Full in φ

...

hZ::T: half height/length of the cone

SolidStateDetectors.ConstructiveSolidGeometry.ConeMantle — Type

struct ConeMantle{T,TR,TP,D} <: AbstractSurfacePrimitive{T}

T: Type of values, e.g. Float64

r::TR:

TR = Real -> Cylinder Mantle (a = b = r)

TR = (Real, Real) -> Cone Mantle (rbot = r[1], rtop = r[2])

TR = ((Real,), (Real,)) -> Elliptical Cylinder Mantle (a = r[1][1], b = r[2][1])

TR = ((Real, Real),(Real, Real)) -> Elliptical Cone Mantle

(ain = r[1][1], aout = r[1][2], bin = r[2][1], bout = r[2][2]) * Not all are implemented yet

φ::TP:

TP = Nothing <-> Full in φ

...

hZ::T: half hight/length of the cone mantle

D: :inwards or :outwards: Whethe the normal points inside or outside

SolidStateDetectors.ConstructiveSolidGeometry.CylindricalPoint — Type

struct CylindricalPoint{T} <: AbstractCoordinatePoint{T, Cylindrical}

r: Radius in meter

φ: Polar angle in radians. φ == 0 <=> Parallel to x-axis of cartesian coordinate system."

z: z-coordinate in meter

SolidStateDetectors.ConstructiveSolidGeometry.CylindricalVector — Type

struct CylindricalVector{T} <: AbstractCoordinateVector{T, Cylindrical}

r: Radius in meter

φ: Polar angle in radians. φ == 0 <=> Parallel to x-axis of cartesian coordinate system."

z: z-coordinate in meter

SolidStateDetectors.ConstructiveSolidGeometry.Ellipse — Type

struct Ellipse{T,TR} <: AbstractLinePrimitive{T}

r::TR:

TR = Real -> Circle (a = b = r)

TR = (Real, Real) -> Circular Annulus (rin = r[1], rout = r[2])

TR = ((Real,), (Real,)) -> Ellipse (a = r[1][1], b = r[2][1])

TR = ((Real, Real),(Real, Real)) -> Elliptical Annulus

(ain = r[1][1], aout = r[1][2], bin = r[2][1], bout = r[2][2]) * Not all are implemented yet

φ::TP:

TP = Nothing <-> Full in φ

...

SolidStateDetectors.ConstructiveSolidGeometry.EllipticalSurface — Type

EllipticalSurface{T,TR,TP} <: AbstractSurfacePrimitive{T}

r::TR:

TR = Real -> Full Circle (a = b = r)

TR = (Real, Real) -> Circular Annulus (rin = r[1], rout = r[2])

TR = ((Real,), (Real,)) -> Full Ellipse (a = r[1][1], b = r[2][1])

TR = ((Real, Real),(Real, Real)) -> Elliptical Annulus

(ain = r[1][1], aout = r[1][2], bin = r[2][1], bout = r[2][2]) * Not all are implemented yet

φ::TP:

TP = Nothing <-> Full in φ

...

SolidStateDetectors.ConstructiveSolidGeometry.extreme_points — Method

extreme_points(es::AbstractPrimitive{T}) where {T}

General implementation for all primitives.

SolidStateDetectors.ConstructiveSolidGeometry.intersection — Method

intersection(cm::ConeMantle{T,Tuple{T,T}}, l::Line{T}) where {T}

The function will always return 2 CartesianPoint's. If the line just touches the mantle, the two points will be the same. If the line does not touch the mantle at all, the two points will have NaN's as there coordinates. If the line crosses the mantle only once, two points will be returned. The two points will be the same point (the intersection). If the line lies inside the mantle and is parallel to it. The same point will be returned which is the origin of the line.

SolidStateDetectors.ConstructiveSolidGeometry.intersection — Method

intersection(cm::EllipsoidMantle{T,T}, l::Line{T}) where {T}

The function will always return 2 CartesianPoint's. If the line just touches the mantle, the two points will be the same. If the line does not touch the mantle at all, the two points will have NaN's as there coordinates.

SolidStateDetectors.ConstructiveSolidGeometry.intersection — Method

intersection(cm::EllipsoidMantle{T,NTuple{3,T}}, l::Line{T}) where {T}

The function will always return 2 CartesianPoint's. If the line just touches the mantle, the two points will be the same. If the line does not touch the mantle at all, the two points will have NaN's as there coordinates.

SolidStateDetectors.ConstructiveSolidGeometry.intersection — Method

intersection(tm::TorusMantle{T}, l::Line{T}) where {T}

The function will always return 4 CartesianPoint's. If the line just touches the mantle, the points will be the same. If the line does not touch the mantle at all, the points will have NaN's as there coordinates.

Solve: solve (sqrt((L1 + λ*D1)^2 + (L2 + λ*D2)^2)-R)^2 + (L3 + λ*D3)^2 = r^2 for λ