API

Modules

SolidStateDetectors.ConstructiveSolidGeometry

Types

SolidStateDetectors.ADLChargeDriftModel

SolidStateDetectors.AbstractChargeDensity

SolidStateDetectors.AbstractConfig

SolidStateDetectors.AbstractImpurityDensity

SolidStateDetectors.ConstantChargeDensity

SolidStateDetectors.ConstantImpurityDensity

SolidStateDetectors.ConstructiveSolidGeometry.Box

SolidStateDetectors.ConstructiveSolidGeometry.CSGDifference

SolidStateDetectors.ConstructiveSolidGeometry.CSGIntersection

SolidStateDetectors.ConstructiveSolidGeometry.CSGUnion

SolidStateDetectors.ConstructiveSolidGeometry.CartesianPoint

SolidStateDetectors.ConstructiveSolidGeometry.CartesianVector

SolidStateDetectors.ConstructiveSolidGeometry.Cone

SolidStateDetectors.ConstructiveSolidGeometry.ConeMantle

SolidStateDetectors.ConstructiveSolidGeometry.CylindricalPoint

SolidStateDetectors.ConstructiveSolidGeometry.CylindricalVector

SolidStateDetectors.ConstructiveSolidGeometry.Ellipse

SolidStateDetectors.ConstructiveSolidGeometry.EllipticalSurface

SolidStateDetectors.Contact

SolidStateDetectors.CylindricalChargeDensity

SolidStateDetectors.CylindricalImpurityDensity

SolidStateDetectors.DiscreteAxis

SolidStateDetectors.DiscreteAxis

SolidStateDetectors.ElectricFieldChargeDriftModel

SolidStateDetectors.ElectricPotential

SolidStateDetectors.ElectricPotential

SolidStateDetectors.Event

SolidStateDetectors.Grid

SolidStateDetectors.LinearChargeDensity

SolidStateDetectors.LinearImpurityDensity

SolidStateDetectors.PointType

SolidStateDetectors.PointTypes

SolidStateDetectors.PointTypes

SolidStateDetectors.PotentialSimulationSetup

SolidStateDetectors.Simulation

SolidStateDetectors.SolidStateDetector

Functions

SolidStateDetectors.ConstructiveSolidGeometry.extreme_points

SolidStateDetectors.ConstructiveSolidGeometry.intersection

SolidStateDetectors.ConstructiveSolidGeometry.intersection

SolidStateDetectors.ConstructiveSolidGeometry.intersection

SolidStateDetectors.ConstructiveSolidGeometry.intersection

SolidStateDetectors.ConstructiveSolidGeometry.roots_of_4th_order_polynomial

SolidStateDetectors.RBArray

SolidStateDetectors.RBExtBy2Array

SolidStateDetectors.RBExtBy2Array

SolidStateDetectors._convert_to_original_potential

SolidStateDetectors._drift_charge!

SolidStateDetectors._get_closed_potential

SolidStateDetectors.add_fano_noise

SolidStateDetectors.apply_initial_state!

SolidStateDetectors.apply_initial_state!

SolidStateDetectors.calculate_capacitance

SolidStateDetectors.calculate_electric_field!

SolidStateDetectors.calculate_electric_potential!

SolidStateDetectors.calculate_weighting_potential!

SolidStateDetectors.get_active_volume

SolidStateDetectors.get_electron_drift_field

SolidStateDetectors.get_hole_drift_field

SolidStateDetectors.get_path_to_example_config_files

SolidStateDetectors.get_rbidx_right_neighbour

SolidStateDetectors.innerloops!

SolidStateDetectors.innerloops!

SolidStateDetectors.nidx

SolidStateDetectors.parse_config_file

SolidStateDetectors.point_type

SolidStateDetectors.readsiggen

SolidStateDetectors.refine!

SolidStateDetectors.refine!

SolidStateDetectors.refine_scalar_potential

SolidStateDetectors.siggentodict

SolidStateDetectors.simulate!

SolidStateDetectors.update!

SolidStateDetectors.update_till_convergence!

SolidStateDetectors.update_till_convergence!

Documentation

SolidStateDetectors.ADLChargeDriftModel — Type

ADLChargeDriftModel{T <: SSDFloat, M <: AbstractDriftMaterial, N, TM <: AbstractTemperatureModel{T}} <: AbstractChargeDriftModel{T}

Fields

electrons::CarrierParameters{T}

holes::CarrierParameters{T}

crystal_orientation::SMatrix{3,3,T,9}

γ::SVector{N,SMatrix{3,3,T,9}}

parameters::ADLParameters{T}

temperaturemodel::TM

SolidStateDetectors.AbstractChargeDensity — Type

# abstract type AbstractChargeDensity{T <: SSDFloat} end

Charge densities <: AbstractChargeDensity should be defined to return (via get_charge_density-method) a charge density in SI units, thus, in Q/m^3.

SolidStateDetectors.AbstractConfig — Type

abstract type AbstractConfig{T <: SSDFloat} end

Supertype of all detector/world/object configs.

User defined geometries must be subtype of AbstractConfig{T}.

There are a few functions which must be defined for a user config, e.g. struct UserConfig{T} <: AbstractConfig{T}:

For cylindrical grids:

in(pt::CylindricalPoint{T}, config::UserConfig{T})::Bool where {T <: SSDFloat}

Grid(config::UserConfig{T})::Grid{T, 3, Cylindrical} where {T <: SSDFloat}

get_ρ_and_ϵ(pt::CylindricalPoint{T}, config::UserConfig{T})::Tuple{T, T} where {T <: SSDFloat}

set_pointtypes_and_fixed_potentials!(pointtypes::Array{PointType, 3}, potential::Array{T, 3}, grid::Grid{T, 3, Cylindrical}, config::UserConfig{T}; weighting_potential_channel_idx::Union{Missing, Int} = missing)::Nothing where {T <: SSDFloat}

For cartesian grids:

in(pt::CartesianPoint{3, T}, config::UserConfig{T})::Bool

Grid(config::UserConfig{T})::Grid{T, 3, Cartesian} where {T <: SSDFloat}

get_ρ_and_ϵ(pt::CartesianPoint{3, T}, config::UserConfig{T})::Tuple{T, T} where {T <: SSDFloat}

set_pointtypes_and_fixed_potentials!(pointtypes::Array{PointType, 3}, potential::Array{T, 3}, grid::Grid{T, 3, Cartesian}, config::UserConfig{T}; weighting_potential_channel_idx::Union{Missing, Int} = missing)::Nothing where {T <: SSDFloat}

SolidStateDetectors.AbstractImpurityDensity — Type

# abstract type AbstractImpurityDensity{T <: SSDFloat} end

Impurity densities <: AbstractImpurityDensity should be defined to return (via get_impurity_density-method) a density of impurities in SI units, thus, in 1/m^3. For semiconductors, this will be converted internally into a charge distribution by multiplication with the elementary charge. The sign of the impurity density is important. It is taken into account in the conversion to a charge density and, thus, defines where the semiconductor is n-type or p-type.

SolidStateDetectors.ConstantChargeDensity — Type

struct ConstantChargeDensity{T <: SSDFloat} <: AbstractChargeDensity{T}

Returns always a fixed charge density.

SolidStateDetectors.ConstantImpurityDensity — Type

struct ConstantImpurityDensity{T <: SSDFloat} <: AbstractImpurityDensity{T}

Returns always a fixed Impurity density.

SolidStateDetectors.Contact — Type

mutable struct Contact{T} <: AbstractContact{T}

T: Type of precision.

SolidStateDetectors.CylindricalChargeDensity — Type

struct CylindricalChargeDensity{T <: SSDFloat} <: AbstractChargeDensity{T}

Simple charge density model which assumes a linear gradient in charge density in each spatial dimension of a cylindrical coordinate system. offsets::NTuple{3, T} are the charge densities at 0 and gradients::NTuple{3, T} are the linear slopes in r and z direction.

SolidStateDetectors.CylindricalImpurityDensity — Type

struct CylindricalImpurityDensity{T <: SSDFloat} <: AbstractImpurityDensity{T}

Simple Impurity density model which assumes a linear gradient in Impurity density in each spatial dimension of a cylindrical coordinate system. offsets::NTuple{3, T} are the Impurity densities at 0 and gradients::NTuple{3, T} are the linear slopes in r and z direction.

SolidStateDetectors.DiscreteAxis — Type

DiscreteAxis{T, BL, BR} <: AbstractAxis{T, BL, BR}

T: Type of ticks

BL, BR ∈ {:periodic, :reflecting, :infinite, :r0, :fixed}

BL: left boundary condition

BR: right boundary condition

I: IntervalSets.Interval (closed or open boundaries)

SolidStateDetectors.DiscreteAxis — Method

DiscreteAxis(left_endpoint::T, right_endpoint::T, BL::Symbol, BR::Symbol, L::Symbol, R::Symbol, ticks::AbstractVector{T}) where {T}

T: Type of ticks

BL, BR ∈ {:periodic, :reflecting, :infinite, :r0, :fixed}

L, R {:closed, :open}

ticks: Ticks of the axis

SolidStateDetectors.ElectricFieldChargeDriftModel — Type

ElectricFieldChargeDriftModel <: AbstractChargeDriftModel

SolidStateDetectors.ElectricPotential — Method

ElectricPotential(setup::PotentialSimulationSetup{T, 3, Cartesian} ; kwargs...)::ElectricPotential{T, 3, Cartesian}

Extracts the electric potential from setup.

SolidStateDetectors.ElectricPotential — Method

ElectricPotential(setup::PotentialSimulationSetup{T, 3, Cylindrical} ; kwargs...)::ElectricPotential{T, 3, Cylindrical}

Extracts the electric potential from setup and extrapolate it to an 2π grid.

For 2D grids (r and z) the user has to set the keyword n_points_in_φ::Int, e.g.: n_points_in_φ = 36.

SolidStateDetectors.Event — Type

mutable struct Event{T <: SSDFloat}

Collection struct for individual events. This (mutable) struct is ment to be used to look at individual events, not to process a huge amount of events.

SolidStateDetectors.Grid — Type

T: tick type
N: N dimensional
S: System (Cartesian, Cylindrical...)

SolidStateDetectors.LinearChargeDensity — Type

struct LinearChargeDensity{T <: SSDFloat} <: AbstractChargeDensity{T}

Simple charge density model which assumes a linear gradient in charge density in each dimension of a Cartesian coordinate system. offsets::NTuple{3, T} are the charge densities at 0 and gradients::NTuple{3, T} are the linear slopes in each dimension, x, y and z.

SolidStateDetectors.LinearImpurityDensity — Type

struct LinearImpurityDensity{T <: SSDFloat} <: AbstractImpurityDensity{T}

Simple Impurity density model which assumes a linear gradient in Impurity density in each dimension of a Cartesian coordinate system. offsets::NTuple{3, T} are the Impurity densities at 0 and gradients::NTuple{3, T} are the linear slopes in each dimension, x, y and z.

SolidStateDetectors.PointType — Type

const PointType = UInt8

Stores certain information about a grid point via bit-flags.

Right now there are:

`const update_bit      = 0x01`
`const undepleted_bit  = 0x02`
`const pn_junction_bit = 0x04`

How to get information out of a PointType variable pt:

pt & update_bit == 0 -> do not update this point (for fixed points)

pt & update_bit > 0 -> do update this point

pt & undepleted_bit > 0 -> this point is undepleted

pt & pn_junction_bit > 0 -> this point belongs to the solid state detector. So it is in the volume of the n-type or p-type material.

SolidStateDetectors.PointTypes — Type

PointTypes{T, N, S} <: AbstractArray{T, N}

PointTypes stores the point type of each grid point.

SolidStateDetectors.PointTypes — Method

PointTypes(setup::PotentialSimulationSetup{T, 3, Cylindrical} ; kwargs...)::PointTypes{T, 3, Cylindrical}

Extracts the point types from setup and extrapolates it to an 2π grid.

For 2D grids (r and z) the user has to set the keyword n_points_in_φ::Int, e.g.: n_points_in_φ = 36.

SolidStateDetectors.PotentialSimulationSetup — Type

PotentialSimulationSetup{T, N, S} <: AbstractPotentialSimulationSetup{T, N}

Collection struct. It holds the grid, the potential, the point types, the charge density and the dielectric distribution.

SolidStateDetectors.Simulation — Type

mutable struct Simulation{T <: SSDFloat, CS <: AbstractCoordinateSystem} <: AbstractSimulation{T}

Collection of all parts of a Simulation of a Solid State Detector.

SolidStateDetectors.SolidStateDetector — Type

mutable struct SolidStateDetector{T <: SSDFloat} <: AbstractConfig{T}

SolidStateDetectors.RBArray — Method

RBExtBy2Array( et::Type, g::Grid{T, N, Cylindrical} )::Array{et, N + 1} where {T, N}

Returns a RedBlack array for the grid g.

SolidStateDetectors.RBExtBy2Array — Method

RBExtBy2Array( et::Type, g::Grid{T, N, Cylindrical} )::Array{et, N + 1} where {T, N}

Returns a RedBlack array for the grid g. The RedBlack array is extended in its size by 2 in each geometrical dimension.

SolidStateDetectors.RBExtBy2Array — Method

RBExtBy2Array( et::Type, g::Grid{T, 3, Cartesian} )::Array{et, 4} where {T}

Returns a RedBlack array for the grid g. The RedBlack array is extended in its size by 2 in each geometrical dimension.

SolidStateDetectors._convert_to_original_potential — Method

_convert_to_original_potential(::Type{P}, data, grid) where {P, T, CS}

Basically the counterpart to _get_closed_potential.

SolidStateDetectors._drift_charge! — Method

_drift_charge!(...)

Before calling this function one should check that startpos is inside det: in(startpos, det)

SolidStateDetectors._get_closed_potential — Method

_get_closed_potential(p::ScalarPotential{T,3,CS}) where {T, CS}

Returns an closed Grid & Potential: E.g. if one of the axis is {:closed,:open} it will turn this into {:closed,:closed} and also extend the data field of the potential in the respective dimension and fill it with the respective values.

SolidStateDetectors.add_fano_noise — Function

add_fano_noise(E_dep::RealQuantity, E_ionisation::RealQuantity, f_fano::Real)::RealQuantity

Add Fano noise to an energy deposition E_dep, assuming a detector material ionisation energy E_ionisation and a Fano factor f_fano.

SolidStateDetectors.apply_initial_state! — Method

function apply_initial_state!(sim::Simulation{T}, ::Type{ElectricPotential}, grid::Grid{T} = Grid(sim))::Nothing

Applies the initial state of the electric potential calculation. It overwrites sim.electric_potential, sim.q_eff_imp, sim.q_eff_fix, sim.ϵ and sim.point_types.

SolidStateDetectors.apply_initial_state! — Method

function apply_initial_state!(sim::Simulation{T}, ::Type{WeightingPotential}, contact_id::Int, grid::Grid{T} = Grid(sim))::Nothing

Applies the initial state of the weighting potential calculation for the contact with the id contact_id. It overwrites sim.weighting_potentials[contact_id].

SolidStateDetectors.calculate_capacitance — Method

calculate_capacitance(sim::Simulation{T})::T where {T <: SSDFloat}

Calculates the capacitance of an detector in Farad.

SolidStateDetectors.calculate_electric_field! — Method

calculate_electric_field!(sim::Simulation{T}, args...; n_points_in_φ::Union{Missing, Int} = missing, kwargs...)::Nothing

ToDo...

SolidStateDetectors.calculate_electric_potential! — Method

calculate_electric_potential!(sim::Simulation{T}; kwargs...)::Nothing

Compute the electric potential for the given Simulation sim on an adaptive grid through successive over relaxation.

There are several <keyword arguments> which can be used to tune the computation:

Keywords

convergence_limit::Real: convergence_limit times the bias voltage sets the convergence limit of the relaxation. The convergence value is the absolute maximum difference of the potential between two iterations of all grid points. Default of convergence_limit is 2e-6 (times bias voltage).

refinement_limits: Defines the maximum relative (to applied bias voltage) allowed differences of the potential value of neighbored grid points in each dimension for each refinement.

rl::Real -> One refinement with rl equal in all 3 dimensions

rl::Tuple{<:Real,<:Real,<:Real} -> One refinement with rl set individual for each dimension

rl::Vector{<:Real} -> length(l) refinements with rl[i] being the limit for the i-th refinement.

rl::Vector{<:Real,<:Real,<:Real}} -> length(rl) refinements with rl[i] being the limits for the i-th refinement.

min_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the minimum allowed distance between two grid ticks for each dimension. It prevents the refinement to make the grid to fine. Default is 1e-5 for linear axes and 0.25 * r_max for the polar axis in case of a cylindrical grid.

max_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the maximum allowed distance between two grid ticks for each dimension used in the initialization of the grid. Default is 1/4 of size of the world of the respective dimension.

grid::Grid: Initial grid used to start the simulation. Default is Grid(sim).

max_distance_ratio::Real: Maximum allowed ratio between the two distances in any dimension to the two neighbouring grid points. If the ratio is too large, additional ticks are generated such that the new ratios are smaller than max_distance_ratio. Default is 5.

depletion_handling::Bool: Enables the handling of undepleted regions. Default is false.

use_nthreads::Int: Number of threads to use in the computation. Default is Base.Threads.nthreads(). The environment variable JULIA_NUM_THREADS must be set appropriately before the Julia session was started (e.g. export JULIA_NUM_THREADS=8 in case of bash).

sor_consts::Union{<:Real, NTuple{2, <:Real}}: Two element tuple in case of cylindrical coordinates. First element contains the SOR constant for r = 0. Second contains the constant at the outer most grid point in r. A linear scaling is applied in between. First element should be smaller than the second one and both should be ∈ [1.0, 2.0]. Default is [1.4, 1.85]. In case of cartesian coordinates only one value is taken.

max_n_iterations::Int: Set the maximum number of iterations which are performed after each grid refinement. Default is 10000. If set to -1 there will be no limit.

verbose::Bool=true: Boolean whether info output is produced or not.

SolidStateDetectors.calculate_weighting_potential! — Method

calculate_weighting_potential!(sim::Simulation{T}, contact_id::Int; kwargs...)::Nothing

Compute the weighting potential for the contact with id contact_id for the given Simulation sim on an adaptive grid through successive over relaxation.

There are several <keyword arguments> which can be used to tune the computation:

Keywords

convergence_limit::Real: convergence_limit times the bias voltage sets the convergence limit of the relaxation. The convergence value is the absolute maximum difference of the potential between two iterations of all grid points. Default of convergence_limit is 2e-6 (times bias voltage).

refinement_limits: Defines the maximum relative (to applied bias voltage) allowed differences of the potential value of neighbored grid points in each dimension for each refinement.

rl::Real -> One refinement with rl equal in all 3 dimensions

rl::Tuple{<:Real,<:Real,<:Real} -> One refinement with rl set individual for each dimension

rl::Vector{<:Real} -> length(l) refinements with rl[i] being the limit for the i-th refinement.

rl::Vector{<:Real,<:Real,<:Real}} -> length(rl) refinements with rl[i] being the limits for the i-th refinement.

min_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the minimum allowed distance between two grid ticks for each dimension. It prevents the refinement to make the grid to fine. Default is 1e-5 for linear axes and 0.25 * r_max for the polar axis in case of a cylindrical grid.

max_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the maximum allowed distance between two grid ticks for each dimension used in the initialization of the grid. Default is 1/4 of size of the world of the respective dimension.

grid::Grid: Initial grid used to start the simulation. Default is Grid(sim).

max_distance_ratio::Real: Maximum allowed ratio between the two distances in any dimension to the two neighbouring grid points. If the ratio is too large, additional ticks are generated such that the new ratios are smaller than max_distance_ratio. Default is 5.

depletion_handling::Bool: Enables the handling of undepleted regions. Default is false.

use_nthreads::Int: Number of threads to use in the computation. Default is Base.Threads.nthreads(). The environment variable JULIA_NUM_THREADS must be set appropriately before the Julia session was started (e.g. export JULIA_NUM_THREADS=8 in case of bash).

sor_consts::Union{<:Real, NTuple{2, <:Real}}: Two element tuple in case of cylindrical coordinates. First element contains the SOR constant for r = 0. Second contains the constant at the outer most grid point in r. A linear scaling is applied in between. First element should be smaller than the second one and both should be ∈ [1.0, 2.0]. Default is [1.4, 1.85]. In case of cartesian coordinates only one value is taken.

max_n_iterations::Int: Set the maximum number of iterations which are performed after each grid refinement. Default is 10000. If set to -1 there will be no limit.

verbose::Bool=true: Boolean whether info output is produced or not.

SolidStateDetectors.get_active_volume — Method

get_active_volume(pts::PointTypes{T}) where {T}

Returns an approximation of the active volume of the detector by summing up the cell volumes of all depleted cells.

SolidStateDetectors.get_electron_drift_field — Method

get_electron_drift_field(ef::Array{SVector{3, T},3}, chargedriftmodel::AbstractChargeDriftModel)::Array{SVector{3,T},3} where {T <: SSDFloat}

Applies the charge drift model onto the electric field vectors. The field vectors have to be in cartesian coordinates.

SolidStateDetectors.get_hole_drift_field — Method

get_hole_drift_field(ef::Array{SVector{3, T},3}, chargedriftmodel::AbstractChargeDriftModel)::Array{SVector{3,T},3} where {T <: SSDFloat}

Applies the charge drift model onto the hole field vectors. The field vectors have to be in cartesian coordinates.

SolidStateDetectors.get_path_to_example_config_files — Method

get_path_to_example_config_files()::String

Returns the path to example detector config files provided by the package.

SolidStateDetectors.get_rbidx_right_neighbour — Method

get_rbidx_right_neighbour(rbidx::Int, ::Val{true}, ::Val{true})::Int

needs docu...

SolidStateDetectors.innerloops! — Method

innerloops!(  iz::Int, rb_tar_idx::Int, rb_src_idx::Int, gw_x::Array{T, 2}, gw_y::Array{T, 2}, gw_z::Array{T, 2}, fssrb::PotentialSimulationSetupRB{T, 3, 4, Cartesian},
                            update_even_points::Val{even_points},
                            depletion_handling::Val{depletion_handling_enabled},
                            )::Nothing where {T, even_points, depletion_handling_enabled}

(Vectorized) inner loop for Cartesian coordinates. This function does all the work in the field calculation.

SolidStateDetectors.innerloops! — Method

innerloops!(  ir::Int, rb_tar_idx::Int, rb_src_idx::Int, gw_r::Array{T, 2}, gw_φ::Array{T, 2}, gw_z::Array{T, 2}, fssrb::PotentialSimulationSetupRB{T, 3, 4, Cylindrical},
                            update_even_points::Val{even_points},
                            depletion_handling::Val{depletion_handling_enabled},
                        )::Nothing where {T, even_points, depletion_handling_enabled}

(Vectorized) inner loop for Cylindrical coordinates. This function does all the work in the field calculation.

SolidStateDetectors.nidx — Method

nidx( rbidx::Int, ::Val{true}, ::Val{true})::Int

first type argument: type of the orgal point (for even points -> Val{true}(), else Val{false}()) second type argument: is sum of other point indices even or odd -> (if sum is even -> Val{true}(), else Val{false}())

SolidStateDetectors.parse_config_file — Method

parse_config_file(filename::AbstractString)::SolidStateDetector{T} where {T <: SSDFloat}

Reads in a config file and returns an Detector struct which holds all information specified in the config file. Currently supported formats for the config file: .json, .yaml

SolidStateDetectors.point_type — Method

For charge drift...

SolidStateDetectors.readsiggen — Method

readsiggen(file_path::String[, T::Type=Float64])

Read the '*.config' file in 'file_path' for SigGen and returns a dictionary of all parameters. Non-existing parameteres are set to 0. ...

Arguments

file_path::String: file path for the SigGen config file.

T::Type=Float64: type of the parameters in the output dictionary.

...

SolidStateDetectors.refine! — Method

function refine!(sim::Simulation{T}, ::Type{ElectricPotential}, max_diffs::Tuple{<:Real,<:Real,<:Real}, minimum_distances::Tuple{<:Real,<:Real,<:Real})

Takes the current state of sim.electric_potential and refines it with respect to the input arguments max_diffs and minimum_distances.

SolidStateDetectors.refine! — Method

function refine!(sim::Simulation{T}, ::Type{WeightingPotential}, max_diffs::Tuple{<:Real,<:Real,<:Real}, minimum_distances::Tuple{<:Real,<:Real,<:Real})

Takes the current state of sim.weighting_potentials[contact_id] and refines it with respect to the input arguments max_diffs and minimum_distances.

SolidStateDetectors.refine_scalar_potential — Method

refine_scalar_potential(p::ScalarPotential{T}, max_diffs::NTuple{3, T}, minimum_distances::NTuple{3, T}; 
    only2d::Val{only_2d} = Val(size(p.data, 2)==1)) where {T, only_2d}

Refine any scalar potential p.

Extent the grid to be a closed grid in all dimensions.

Refine the axis of the grid based on max_diffs and minimum_applied_potential: Insert N new ticks between to existing ticks such that the potential difference between each tick becomes smaller than max_diff[i] (i -> dimension) but that the distances between the ticks stays larger than minimum_distances[i].

Create the new data array for the refined grid and fill it by interpolation of the the initial (coarse) grid.

SolidStateDetectors.siggentodict — Method

siggentodict(config::Dict[, units::Dict, detector_name::String])

Converts the dictionary containing the parameters from a SigGen config file to a SSD config dictionary. This dictionary can be saved as a JSON file using the JSON package and 'JSON.print(file, config, 4)'. The 'detector_name' is set to "Public Inverted Coax" by default to inherit the colour scheme. ...

Arguments

config::Dict: dictionary containing SigGen parameters (output of readsiggen()).

units::Dict: units used in SigGen file (set to 'mm', 'deg', 'V' and 'K').

detector_name::String: name of the detector.

...

SolidStateDetectors.simulate! — Method

function simulate!( sim::Simulation{T};  
                refinement_limits = [0.2, 0.1, 0.05],
                min_tick_distance::Union{Missing, length_unit, Tuple{length_unit, angle_unit, length_unit}} = missing,
                max_tick_distance::Union{Missing, length_unit, Tuple{length_unit, angle_unit, length_unit}} = missing,
                max_distance_ratio::Real = 5,
                verbose::Bool = false,
                use_nthreads::Int = Base.Threads.nthreads(),
                sor_consts::Union{Missing, <:Real, Tuple{<:Real,<:Real}} = missing,
                max_n_iterations::Int = -1,
                depletion_handling::Bool = false, 
                convergence_limit::Real = 1e-7 ) where {T <: SSDFloat}

Keywords

convergence_limit::Real: convergence_limit times the bias voltage sets the convergence limit of the relaxation. The convergence value is the absolute maximum difference of the potential between two iterations of all grid points. Default of convergence_limit is 2e-6 (times bias voltage).

refinement_limits: Defines the maximum relative (to applied bias voltage) allowed differences of the potential value of neighbored grid points in each dimension for each refinement.

rl::Real -> One refinement with rl equal in all 3 dimensions

rl::Tuple{<:Real,<:Real,<:Real} -> One refinement with rl set individual for each dimension

rl::Vector{<:Real} -> length(l) refinements with rl[i] being the limit for the i-th refinement.

rl::Vector{<:Real,<:Real,<:Real}} -> length(rl) refinements with rl[i] being the limits for the i-th refinement.

min_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the minimum allowed distance between two grid ticks for each dimension. It prevents the refinement to make the grid to fine. Default is 1e-5 for linear axes and 0.25 * r_max for the polar axis in case of a cylindrical grid.

max_tick_distance::Tuple{<:Quantity, <:Quantity, <:Quantity}: Tuple of the maximum allowed distance between two grid ticks for each dimension used in the initialization of the grid. Default is 1/4 of size of the world of the respective dimension.

max_distance_ratio::Real: Maximum allowed ratio between the two distances in any dimension to the two neighbouring grid points. If the ratio is too large, additional ticks are generated such that the new ratios are smaller than max_distance_ratio. Default is 5.

depletion_handling::Bool: Enables the handling of undepleted regions. Default is false.

use_nthreads::Int: Number of threads to use in the computation. Default is Base.Threads.nthreads(). The environment variable JULIA_NUM_THREADS must be set appropriately before the Julia session was started (e.g. export JULIA_NUM_THREADS=8 in case of bash).

sor_consts::Union{<:Real, NTuple{2, <:Real}}: Two element tuple in case of cylindrical coordinates. First element contains the SOR constant for r = 0. Second contains the constant at the outer most grid point in r. A linear scaling is applied in between. First element should be smaller than the second one and both should be ∈ [1.0, 2.0]. Default is [1.4, 1.85]. In case of cartesian coordinates only one value is taken.

max_n_iterations::Int: Set the maximum number of iterations which are performed after each grid refinement. Default is -1. If set to -1 there will be no limit.

verbose::Bool=true: Boolean whether info output is produced or not.

SolidStateDetectors.update! — Method

update!(fssrb::PotentialSimulationSetupRB{T, 3, 4, S}, RBT::DataType)::Nothing

Loop over even grid points. A point is even if the sum of its cartesian indicies (of the not extended grid) is even. Even points get the red black index (rbi) = 2. ( -> rbpotential[ inds..., rbi ]).

SolidStateDetectors.update_till_convergence! — Method

function update_till_convergence!( sim::Simulation{T} ::Type{ElectricPotential}, convergence_limit::Real; kwargs...)::T

Takes the current state of sim.electric_potential and updates it until it has converged.

SolidStateDetectors.update_till_convergence! — Method

function update_till_convergence!( sim::Simulation{T} ::Type{WeightingPotential}, contact_id::Int, convergence_limit::Real; kwargs...)::T

Takes the current state of sim.weighting_potentials[contact_id] and updates it until it has converged.

SolidStateDetectors.ConstructiveSolidGeometry — Module

module ConstructiveSolidGeometry

SolidStateDetectors.ConstructiveSolidGeometry.Box — Type

Box{T, CO} <: AbstractVolumePrimitive{T}

T: Type of values, e.g. Float64 CO: ClosedPrimitive or OpenPrimitive <-> whether surface belongs to it or not

SolidStateDetectors.ConstructiveSolidGeometry.CSGDifference — Type

struct CSGDifference{T, A <: AbstractGeometry{T}, B <: AbstractGeometry{T}} <: AbstractConstructiveGeometry{T}

a && !b

SolidStateDetectors.ConstructiveSolidGeometry.CSGIntersection — Type

struct CSGIntersection{T, A <: AbstractGeometry{T}, B <: AbstractGeometry{T}} <: AbstractConstructiveGeometry{T}

a && b

SolidStateDetectors.ConstructiveSolidGeometry.CSGUnion — Type

struct CSGUnion{T, A <: AbstractGeometry{T}, B <: AbstractGeometry{T}} <: AbstractConstructiveGeometry{T}

a || b

SolidStateDetectors.ConstructiveSolidGeometry.CartesianPoint — Type

struct CartesianPoint{T} <: AbstractCoordinatePoint{T, Cartesian}

x: x-coordinate in meter

y: y-coordinate in meter

z: z-coordinate in meter

SolidStateDetectors.ConstructiveSolidGeometry.CartesianVector — Type

struct CartesianVector{T} <: AbstractCoordinateVector{T, Cartesian}

x: x-component in meter

y: y-component in meter

z: z-component in meter

SolidStateDetectors.ConstructiveSolidGeometry.Cone — Type

struct Cone{T,CO,TR,TP} <: AbstractVolumePrimitive{T, CO}

T: Type of values, e.g. Float64 CO: ClosedPrimitive or OpenPrimitive <-> whether surface belongs to it or not

r::TR:

TR = Real -> Cylinder

TR = (Real, Real) -> Tube (rin = r[1], rout = r[2])

TR = ((Real,), (Real,)) Solid widening Cylinder -> (rbot = r[1][1], rtop = r[1][2])

TR = ((Real,Real), (Real,Real)) Solid widening Tube ->

(rbotin = r[1][1], rbotout = r[1][2], rtopin = r[2][1], rtopout = r[2][2]) * TR = (Nothing, (Real,Real)) Cone -> (rbotin = rbotout = 0, rtopin = r[2][1], rtopout = r[2][2]) * TR = ((Real,Real), Nothing) Cone -> (rbotin = r[1][1], rbotout = r[1][2], rtopin = rtopout = 0) * ... (+ elliptical cases -> (a, b)) * Not all are implemented yet

φ::TP:

TP = Nothing <-> Full in φ

...

hZ::T: half height/length of the cone

SolidStateDetectors.ConstructiveSolidGeometry.ConeMantle — Type

struct ConeMantle{T,TR,TP,D} <: AbstractSurfacePrimitive{T}

T: Type of values, e.g. Float64

r::TR:

TR = Real -> Cylinder Mantle (a = b = r)

TR = (Real, Real) -> Cone Mantle (rbot = r[1], rtop = r[2])

TR = ((Real,), (Real,)) -> Elliptical Cylinder Mantle (a = r[1][1], b = r[2][1])

TR = ((Real, Real),(Real, Real)) -> Elliptical Cone Mantle

(ain = r[1][1], aout = r[1][2], bin = r[2][1], bout = r[2][2]) * Not all are implemented yet

φ::TP:

TP = Nothing <-> Full in φ

...

hZ::T: half hight/length of the cone mantle

D: :inwards or :outwards: Whethe the normal points inside or outside

SolidStateDetectors.ConstructiveSolidGeometry.CylindricalPoint — Type

struct CylindricalPoint{T} <: AbstractCoordinatePoint{T, Cylindrical}

r: Radius in meter

φ: Polar angle in radians. φ == 0 <=> Parallel to x-axis of cartesian coordinate system."

z: z-coordinate in meter

SolidStateDetectors.ConstructiveSolidGeometry.CylindricalVector — Type

struct CylindricalVector{T} <: AbstractCoordinateVector{T, Cylindrical}

r: Radius in meter

φ: Polar angle in radians. φ == 0 <=> Parallel to x-axis of cartesian coordinate system."

z: z-coordinate in meter

SolidStateDetectors.ConstructiveSolidGeometry.Ellipse — Type

struct Ellipse{T,TR} <: AbstractLinePrimitive{T}

r::TR:

TR = Real -> Circle (a = b = r)

TR = (Real, Real) -> Circular Annulus (rin = r[1], rout = r[2])

TR = ((Real,), (Real,)) -> Ellipse (a = r[1][1], b = r[2][1])

TR = ((Real, Real),(Real, Real)) -> Elliptical Annulus

(ain = r[1][1], aout = r[1][2], bin = r[2][1], bout = r[2][2]) * Not all are implemented yet

φ::TP:

TP = Nothing <-> Full in φ

...

SolidStateDetectors.ConstructiveSolidGeometry.EllipticalSurface — Type

EllipticalSurface{T,TR,TP} <: AbstractSurfacePrimitive{T}

r::TR:

TR = Real -> Full Circle (a = b = r)

TR = (Real, Real) -> Circular Annulus (rin = r[1], rout = r[2])

TR = ((Real,), (Real,)) -> Full Ellipse (a = r[1][1], b = r[2][1])

TR = ((Real, Real),(Real, Real)) -> Elliptical Annulus

(ain = r[1][1], aout = r[1][2], bin = r[2][1], bout = r[2][2]) * Not all are implemented yet

φ::TP:

TP = Nothing <-> Full in φ

...

SolidStateDetectors.ConstructiveSolidGeometry.extreme_points — Method

extreme_points(es::AbstractPrimitive{T}) where {T}

General implementation for all primitives.

SolidStateDetectors.ConstructiveSolidGeometry.intersection — Method

intersection(cm::ConeMantle{T,Tuple{T,T}}, l::Line{T}) where {T}

The function will always return 2 CartesianPoint's. If the line just touches the mantle, the two points will be the same. If the line does not touch the mantle at all, the two points will have NaN's as there coordinates. If the line crosses the mantle only once, two points will be returned. The two points will be the same point (the intersection). If the line lies inside the mantle and is parallel to it. The same point will be returned which is the origin of the line.

SolidStateDetectors.ConstructiveSolidGeometry.intersection — Method

intersection(cm::EllipsoidMantle{T,T}, l::Line{T}) where {T}

The function will always return 2 CartesianPoint's. If the line just touches the mantle, the two points will be the same. If the line does not touch the mantle at all, the two points will have NaN's as there coordinates.

SolidStateDetectors.ConstructiveSolidGeometry.intersection — Method

intersection(cm::EllipsoidMantle{T,NTuple{3,T}}, l::Line{T}) where {T}

The function will always return 2 CartesianPoint's. If the line just touches the mantle, the two points will be the same. If the line does not touch the mantle at all, the two points will have NaN's as there coordinates.

SolidStateDetectors.ConstructiveSolidGeometry.intersection — Method

intersection(tm::TorusMantle{T}, l::Line{T}) where {T}

The function will always return 4 CartesianPoint's. If the line just touches the mantle, the points will be the same. If the line does not touch the mantle at all, the points will have NaN's as there coordinates.

Solve: solve (sqrt((L1 + λ*D1)^2 + (L2 + λ*D2)^2)-R)^2 + (L3 + λ*D3)^2 = r^2 for λ

SolidStateDetectors.ConstructiveSolidGeometry.roots_of_4th_order_polynomial — Method

roots_of_4th_order_polynomial(a::T, b::T, c::T, d::T, e::T)

Calculate the 4 (possible) roots of x^4 + ax^3 + bx^2 + cx + d = 0